Support for new source file names in the ADOM components.
Dieter Köhler will prefix his files with dk, in the freely downloadable version of ADOM. And the ADOM shipped as part of Delphi's RTL will have the old, non-prefixed source file names. So by prefixing my adomxmldom.pas with gtc, both the RTL version and the latest version of ADOM and this wrapper can coexist.